Dungeon Unzipping
Current setup is that the zip file must contain directories correctly
labelling the types of segments it holds, based on the variable names in dungeon.
Zip it up in a folder with the dungeon name, and place in plugins/AEGEUS/dungeons/zips
Done!
The schematics are not loaded yet, and debugging features are still in the code.