Easy inventory building(readme for unit testing)
Creating inventory-based menus just got easier.
To create a new inventory menu, create a new InventoryBuilder object, with user who's going to view, size in multiple of 9, and the title.
Then, use setItem(slot, stack) to set the certain slot to the itemstack.
Use setOnClick(slot, inventoryhandler anonymous class) to set what's going to happen when the user clicks on a slot with an item.
Use show(player) to show the player the inventory and register it, to the inventory click handling system.
An example is used in CommandShowSpawners.  Will be removed in the next push, or two, or three maybe.

A solution for unit testing not compiling has been found, even though it is a workaround.
Create a new run configuration based off of maven install by right clicking maven install in lifecycles in the maven tab and create new maven install.  In the command line, set it to "install -DskipTests".
At the bottom of the window, Add Another configuration Junit "Test all in game".
Also add configuration build, and put it above the junit test all configuration, such that it builds the project before running junit tests.  Now when you run this configuration, it should build the project, run junit tests via intellij, then do a maven install, and skipping maven's unit tests.

If you want to run the maven version of unit tests, remove junit test all Configuration and remove -DskipTests.
I personally like the IntelliJ unit testing window a lot better than the text-only maven version.